Moved general ray tracing routines to ray_tracing_general.py
